requirements.txt
asgiref==3.8.1
attrs==23.2.0
Authlib==1.3.1
black==24.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Django==5.0.4
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.0
requests==2.32.3
rpds-py==0.18.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
